Comment a bit more stuff

+++ TekkenTools/GameInteractions/Games.hpp
@@ -19,7 +19,7 @@ enum GameFlag
 	GameCamera = (1 << 2),
 };
 
-// Contains a unique ID for every supported game. Every list should respect the index order here
+// Contains a unique ID for every supported game. GameInfo list should respect the index order here
 enum GameId
 {
 	GameId_t7 = 0,
@@ -29,10 +29,14 @@ enum GameId
 
 struct GameInfo
 {
+	// Displayed name of the game
 	const char* name;
+	// Process that will be opened and main module name
 	const char* processName;
+	// Amount of characters at once the game is able to load
 	uint8_t characterCount;
-	uint8_t flags;
+	// Determines what can be done with the same. See GameFlag enum for more details
+	uint16_t flags;
 };
 
 namespace Games
@@ -43,7 +47,9 @@ namespace Games
 	// Returns the amount of supported games
 	DLLCONTENT size_t GetGamesCount();
 
+	// Istantiate a new game-dependant extractor
 	DLLCONTENT Extractor* FactoryGetExtractor(uint8_t gameId, GameProcess* process, GameData* game);
 
+	// Istantiate a new game-dependant importer
 	DLLCONTENT Importer* FactoryGetImporter(uint8_t gameId, GameProcess* process, GameData* game);
 }
